moved de_noised_covariance_matrix_results to test_covariance_transformer
